+++ src/main/java/com/book_everywhere/service/BookService.java
@@ -6,8 +6,8 @@ import com.book_everywhere.domain.user.User;
 import com.book_everywhere.domain.user.UserRepository;
 import com.book_everywhere.web.dto.book.BookDto;
 import com.book_everywhere.web.dto.book.BookRespDto;
-import com.book_everywhere.web.dto.exception.customs.CustomErrorCode;
-import com.book_everywhere.web.dto.exception.customs.EntityNotFoundException;
+import com.book_everywhere.web.exception.customs.CustomErrorCode;
+import com.book_everywhere.web.exception.customs.EntityNotFoundException;
 import com.book_everywhere.web.dto.review.ReviewRespDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
