+++ src/main/java/com/book_everywhere/service/DataService.java
@@ -14,8 +14,8 @@ import com.book_everywhere.domain.visit.Visit;
 import com.book_everywhere.domain.visit.VisitRepository;
 import com.book_everywhere.web.dto.AllDataDto;
 import com.book_everywhere.web.dto.book.BookRespDto;
-import com.book_everywhere.web.dto.exception.customs.CustomErrorCode;
-import com.book_everywhere.web.dto.exception.customs.EntityNotFoundException;
+import com.book_everywhere.web.exception.customs.CustomErrorCode;
+import com.book_everywhere.web.exception.customs.EntityNotFoundException;
 import com.book_everywhere.web.dto.pin.PinRespDto;
 import com.book_everywhere.web.dto.tag.TagRespDto;
 import lombok.RequiredArgsConstructor;
