+++ src/main/java/com/book_everywhere/service/PinService.java
@@ -4,8 +4,8 @@ import com.book_everywhere.domain.pin.Pin;
 import com.book_everywhere.domain.pin.PinRepository;
 import com.book_everywhere.domain.tagged.Tagged;
 import com.book_everywhere.domain.tagged.TaggedRepository;
-import com.book_everywhere.web.dto.exception.customs.CustomErrorCode;
-import com.book_everywhere.web.dto.exception.customs.EntityNotFoundException;
+import com.book_everywhere.web.exception.customs.CustomErrorCode;
+import com.book_everywhere.web.exception.customs.EntityNotFoundException;
 import com.book_everywhere.web.dto.pin.PinDto;
 import com.book_everywhere.web.dto.pin.PinRespDto;
 import com.book_everywhere.web.dto.pin.PinWithTagCountRespDto;
